+++ drift_dev/test/writer/table_writer_test.dart
@@ -154,4 +154,38 @@ class Foo extends Table {
       }, result.dartOutputs, result.writer);
     },
   );
+
+  test('can have nullable JSON type for null-skipping converter', () async {
+    // Regression test for https://github.com/simolus3/drift/issues/3436
+    final result = await emulateDriftBuild(
+      inputs: {
+        'a|lib/a.dart': '''
+import 'package:drift/drift.dart';
+
+class TodoItems extends Table {
+  IntColumn get id => integer().autoIncrement()();
+  TextColumn get author => text().nullable().map(Author.typeConverter)();
+}
+
+class Author {
+  const Author({
+    required this.name,
+  });
+
+  final String name;
+
+  static JsonTypeConverter2<Author, String, Object?> typeConverter = TypeConverter.json2(
+    fromJson: (json) => Author(name: (json! as Map<String, Object?>)['name'] as String),
+    toJson: (entries) => {'name': entries.name},
+  );
+}
+''',
+      },
+      modularBuild: true,
+    );
+
+    checkOutputs({
+      'a|lib/a.drift.dart': decodedMatches(isNot(contains('Object??'))),
+    }, result.dartOutputs, result.writer);
+  });
 }